requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
lxml==4.5.1
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.3
pytest-html==2.1.1
pytest-metadata==1.9.0
python-decouple==3.3
requests==2.23.0
selenium==3.141.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.3
zipp==3.1.0
